Cleanup main entry and start switching to loggers
- Switch to argparse for processing arguments
- Remove old command switches
- Setup python logging and handle verbosity

also lazy-load main import to delay for logger

Some of the imports e.g. loading lst_ext print out messages as part of
importing, so this allows the logger to be set up first. Normally this
would be bad practice but at the very beginning of startup is acceptable
as it lets us fast-fail.