With external output, check metadata instead of deleting
Previously files were always deleted to check to see if they had been
saved. This makes it so that the mtime/size is checked beforehand and
the file is counted as altered if these have changed. This makes it a
little more robust to things like opening the image viewer.