Fix launching viewers on windows without shell=True
Looks like shell=True is the only way to get the automatic path lookup
on windows, so manually walk over the PATH to find the executable to
launch instead.

This behaviour is also applied on linux/mac but can't immediately see
any reason why this would be a problem, and in any case is a good first
step towards interchangeable dials distributions.