Use dxtbx.sweep_filenames for get_import_run_string
This should add a little more behavioural robustness - there has been
some massaging to check that the existing behaviour is retained for now,
and some edge cases where this won't behave - but it should be an
improvement in consistency.

The tests are more thorough now, also.

This should as a side effect, fix #46.