Prevent reindex dialog showing during startup
If DUI opened on a restored state on reindex, then the reindex GUI would
open (but in the background, hidden behind the DUI window). It also
looks like it churned open/closed a few times in various scenarios.

This patch attempts to close this loop, preventing open at all when
starting DUI and attempting to only open once in other scenarios. It
also cleans up some - but not all - of the opening logic.

Fixes #166