Revert bad fix "removing leftover code that only makes crashes"
This removed code that was writing a log file when the DIALS command
did not otherwise write one. This is probably required to show the log
when viewing previous nodes.

This "crash" would be better solved by resolving the actual str/byte
TypeError in light of the changes to universal_newlines=True in 12d7eae.

This reverts commit 65761673add54e1fc49d9f0811a49bdc4bdcd97a.